+++ CHANGELOG.md
@@ -13,6 +13,7 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 - Don't subscribe to external resources #470
 - Improve frequency search indexing #473
 - Add HTML importer / bookmarks endpoint #432
+- Allow new `Drive` resources without a parent
 
 ## [v0.32.2] - 2022-06-20